+++ src/body/Stash Markdown.user.js
@@ -1,0 +1,34 @@
+/* global marked */
+
+(function () {
+    'use strict';
+
+    const {
+        stash,
+        Stash,
+        waitForElementId,
+        waitForElementClass,
+        waitForElementByXpath,
+        getElementByXpath,
+        insertAfter,
+        reloadImg,
+    } = unsafeWindow.stash;
+
+    function processMarkdown(el) {
+        el.innerHTML = marked.parse(el.innerHTML);
+    }
+
+    stash.addEventListener('page:tag:any', function () {
+        waitForElementByXpath("//div[contains(@class, 'logo-container')]/p", function (xpath, el) {
+            processMarkdown(el);
+        });
+    });
+
+    stash.addEventListener('page:tags', function () {
+        waitForElementByXpath("//div[contains(@class, 'tag-description')]", function (xpath, el) {
+            for (const node of document.querySelectorAll('.tag-description')) {
+                processMarkdown(node);
+            }
+        });
+    });
+})();