+++ src/header/Stash Markdown.user.js
@@ -1,0 +1,11 @@
+// ==UserScript==
+// @name        Stash Markdown
+// @namespace   %NAMESPACE%
+// @description Adds markdown parsing to tag description fields
+// @version     0.1.0
+// @author      7dJx1qP
+// @match       %MATCHURL%
+// @grant       unsafeWindow
+// @require     %LIBRARYPATH%
+// @require     %FILEPATH%
+// ==/UserScript==